+++ sshd-core/src/main/java/org/apache/sshd/common/kex/KexFactoryManager.java
@@ -30,7 +30,6 @@ import org.apache.sshd.common.cipher.BuiltinCiphers;
 import org.apache.sshd.common.cipher.Cipher;
 import org.apache.sshd.common.compression.BuiltinCompressions;
 import org.apache.sshd.common.compression.Compression;
-import org.apache.sshd.common.keyprovider.KeyPairProviderHolder;
 import org.apache.sshd.common.mac.BuiltinMacs;
 import org.apache.sshd.common.mac.Mac;
 import org.apache.sshd.common.signature.SignatureFactoriesManager;
@@ -41,7 +40,7 @@ import org.apache.sshd.common.util.ValidateUtils;
  * Holds KEX negotiation stage configuration
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public interface KexFactoryManager extends KeyPairProviderHolder, SignatureFactoriesManager {
+public interface KexFactoryManager extends SignatureFactoriesManager {
     /**
      * Retrieve the list of named factories for <code>KeyExchange</code>.
      *