+++ sshd-core/src/main/java/org/apache/sshd/server/global/OpenSshHostKeysHandler.java
@@ -40,6 +40,7 @@ import org.apache.sshd.common.util.ValidateUtils;
 import org.apache.sshd.common.util.buffer.Buffer;
 import org.apache.sshd.common.util.buffer.ByteArrayBuffer;
 import org.apache.sshd.common.util.buffer.keys.BufferPublicKeyParser;
+import org.apache.sshd.server.session.ServerSession;
 
 /**
  * An initial handler for &quot;hostkeys-prove-00@openssh.com&quot; request
@@ -90,10 +91,10 @@ public class OpenSshHostKeysHandler extends AbstractOpenSshHostKeysHandler imple
         // according to the specification there MUST be reply required by the server
         ValidateUtils.checkTrue(wantReply, "No reply required for host keys of %s", session);
         Collection<? extends NamedFactory<Signature>> factories =
-                ValidateUtils.checkNotNullAndNotEmpty(
-                        SignatureFactoriesManager.resolveSignatureFactories(this, session),
-                        "No signature factories available for host keys of session=%s",
-                        session);
+            ValidateUtils.checkNotNullAndNotEmpty(
+                    SignatureFactoriesManager.resolveSignatureFactories(this, session),
+                    "No signature factories available for host keys of session=%s",
+                    session);
         if (log.isDebugEnabled()) {
             log.debug("handleHostKeys({})[want-reply={}] received {} keys - factories={}",
                       session, wantReply, GenericUtils.size(keys), NamedResource.getNames(factories));
@@ -104,7 +105,7 @@ public class OpenSshHostKeysHandler extends AbstractOpenSshHostKeysHandler imple
 
         Buffer buf = new ByteArrayBuffer();
         byte[] sessionId = session.getSessionId();
-        KeyPairProvider kpp = Objects.requireNonNull(session.getKeyPairProvider(), "No server keys provider");
+        KeyPairProvider kpp = Objects.requireNonNull(((ServerSession) session).getKeyPairProvider(), "No server keys provider");
         for (PublicKey k : keys) {
             String keyType = KeyUtils.getKeyType(k);
             Signature verifier = ValidateUtils.checkNotNull(